+++ ors-engine/src/main/java/org/heigit/ors/routing/graphhopper/extensions/storages/builders/RoadAccessRestrictionsGraphStorageBuilder.java
@@ -61,6 +61,8 @@ public class RoadAccessRestrictionsGraphStorageBuilder extends AbstractGraphStor
         restrictedValues.add("customers");
         restrictedValues.add("emergency");
         restrictedValues.add("permissive");
+        restrictedValues.add("delivery");
+        restrictedValues.add("permit");
 
         permissiveValues.add("yes");
         permissiveValues.add("designated");
@@ -76,9 +78,9 @@ public class RoadAccessRestrictionsGraphStorageBuilder extends AbstractGraphStor
      * @return The RoadAccessRestrictionStorage object created as part of the initialisation
      * @throws Exception
      */
-    public GraphExtension init(GraphHopper graphhopper, int profileType) throws Exception {
+    public GraphExtension init(GraphHopper graphhopper, int profileType) throws IllegalStateException {
         if (storage != null)
-            throw new Exception("GraphStorageBuilder has been already initialized.");
+            throw new IllegalStateException("GraphStorageBuilder has been already initialized.");
 
         ExtendedStorageProperties parameters;
         parameters = this.parameters;
@@ -100,9 +102,9 @@ public class RoadAccessRestrictionsGraphStorageBuilder extends AbstractGraphStor
      * @return The RoadAccessRestrictionStorage object created as part of the initialisation
      * @throws Exception
      */
-    public GraphExtension init(GraphHopper graphhopper) throws Exception {
+    public GraphExtension init(GraphHopper graphhopper) throws IllegalStateException {
         if (storage != null)
-            throw new Exception("GraphStorageBuilder has been already initialized.");
+            throw new IllegalStateException("GraphStorageBuilder has been already initialized.");
 
         ExtendedStorageProperties parameters;
         parameters = this.parameters;
@@ -166,6 +168,8 @@ public class RoadAccessRestrictionsGraphStorageBuilder extends AbstractGraphStor
                 restrictions = isAccessAllowed(way, VAL_BICYCLE) ? 0 : getRestrictionType(way, VAL_BICYCLE);
             if (RoutingProfileType.isPedestrian(profileType))
                 restrictions = isAccessAllowed(way, "foot") ? 0 : getRestrictionType(way, "foot");
+            if (profileType == RoutingProfileType.UNKNOWN)
+                restrictions = getRestrictionType(way, VAL_ACCESS);
         }
     }
 
@@ -230,6 +234,7 @@ public class RoadAccessRestrictionsGraphStorageBuilder extends AbstractGraphStor
                 case "permissive" -> res |= AccessRestrictionType.PERMISSIVE;
                 case "delivery" -> res |= AccessRestrictionType.DELIVERY;
                 case "customers" -> res |= AccessRestrictionType.CUSTOMERS;
+                case "permit" -> res |= AccessRestrictionType.PERMIT;
                 default -> {
                 }
             }